@@ -13,6 +13,7 @@ import BGU.Group13B.backend.storePackage.PublicAuctionInfo;
 import BGU.Group13B.service.entity.ReviewService;
 import BGU.Group13B.service.entity.ServiceBasketProduct;
 import BGU.Group13B.service.entity.ServiceProduct;
+import BGU.Group13B.service.info.DiscountAccumulationTreeInfo;
 import BGU.Group13B.service.info.ProductInfo;
 import BGU.Group13B.service.info.StoreInfo;
 import org.springframework.stereotype.Service;
@@ -143,6 +144,23 @@ public class Session implements ISession {
     public void purchaseProposalSubmit(int userId, int storeId, int productId, double proposedPrice, int amount) {
         try {
             market.purchaseProposalSubmit(userId, storeId, productId, proposedPrice, amount);
+        } catch (NoPermissionException e) {
+            throw new RuntimeException(e);
+        }
+    }
+    @Override
+    public void purchaseProposalApprove(int managerId, int storeId, int productId){
+        try {
+            market.purchaseProposalApprove(managerId, storeId, productId);
+        } catch (NoPermissionException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    @Override
+    public void purchaseProposalReject(int storeId, int managerId, int bidId){
+        try {
+            market.purchaseProposalReject(managerId,storeId, bidId);
         } catch (NoPermissionException e) {
             throw new RuntimeException(e);
         }
@@ -1294,7 +1312,7 @@ public class Session implements ISession {
     @Override
     public Response<List<DiscountInfo>> getStoreDiscounts(int storeId, int userId) {
         try {
-            return Response.success(market.getStoreDiscounts(storeId, userId).stream().map(d -> new DiscountInfo(d.toString())).collect(Collectors.toList()));
+            return Response.success(market.getStoreDiscounts(storeId, userId).stream().map(d -> new DiscountInfo(d.getDiscountId(), d.toString())).collect(Collectors.toList()));
         } catch (Exception e) {
             return Response.exception(e);
         }
@@ -1304,7 +1322,7 @@ public class Session implements ISession {
     public Response<DiscountInfo> getDiscount(int storeId, int userId, int discountId) {
         try {
             var discount = market.getDiscount(storeId, userId, discountId);
-            return Response.success(new DiscountInfo(discount.toString()));
+            return Response.success(new DiscountInfo(discountId, discount.toString()));
         } catch (Exception e) {
             return Response.exception(e);
         }
@@ -1354,6 +1372,25 @@ public class Session implements ISession {
     public Response<VoidResponse> addDiscountToADDRoot(int storeId, int userId, int discountId) {
         try {
             market.addDiscountToADDRoot(storeId, userId, discountId);
+            return Response.success(new VoidResponse());
+        } catch (Exception e) {
+            return Response.exception(e);
+        }
+    }
+
+    @Override
+    public Response<DiscountAccumulationTreeInfo> getDiscountAccumulationTree(int storeId, int userId) {
+        try{
+            return Response.success(market.getDiscountAccumulationTree(storeId, userId));
+        } catch (Exception e) {
+            return Response.exception(e);
+        }
+    }
+
+    @Override
+    public Response<VoidResponse> deleteStoreAccumulationTree(int storeId, int userId) {
+        try {
+            market.deleteStoreAccumulationTree(storeId, userId);
             return Response.success(new VoidResponse());
         } catch (Exception e) {
             return Response.exception(e);
