@@ -17,6 +17,7 @@ import BGU.Group13B.backend.storePackage.purchaseBounders.PurchaseExceedsPolicyE
 import BGU.Group13B.service.Response;
 import BGU.Group13B.service.SingletonCollection;
 import BGU.Group13B.service.callbacks.AddToUserCart;
+import BGU.Group13B.service.info.DiscountAccumulationTreeInfo;
 import BGU.Group13B.service.info.ProductInfo;
 import BGU.Group13B.service.info.StoreInfo;
 
@@ -70,15 +71,15 @@ public class Market {
     }
 
     public void addReview(String review, int storeId, int productId, int userId) throws NoPermissionException { //TODO:check get store impl
-        storeRepository.getStore(storeId).addReview(review, userId,productId);
+        storeRepository.getStore(storeId).addReview(review, userId, productId);
     }
 
     public void removeReview(int storeId, int productId, int userId) throws NoPermissionException {
-        storeRepository.getStore(storeId).removeReview(userId,productId);
+        storeRepository.getStore(storeId).removeReview(userId, productId);
     }
 
     public Review getReview(int storeId, int productId, int userId) throws NoPermissionException {
-        return storeRepository.getStore(storeId).getReview(userId,productId);
+        return storeRepository.getStore(storeId).getReview(userId, productId);
     }
 
     public List<Review> getAllReviews(int productId, int storeId) throws NoPermissionException {
@@ -86,10 +87,11 @@ public class Market {
     }
 
     public float getProductScore(int storeId, int productId, int userId) throws NoPermissionException {
-        return storeRepository.getStore(storeId).getProductScore(productId,userId );
+        return storeRepository.getStore(storeId).getProductScore(productId, userId);
     }
+
     public float getProductScoreUser(int storeId, int productId, int userId) throws NoPermissionException {
-        return storeRepository.getStore(storeId).getProductScoreUser(productId,userId );
+        return storeRepository.getStore(storeId).getProductScoreUser(productId, userId);
     }
 
     public void addAndSetProductScore(int storeId, int productId, int userId, int score) throws NoPermissionException {
@@ -169,13 +171,13 @@ public class Market {
     }
 
     public List<ProductInfo> search(String searchWords) {
-            List<ProductInfo> products = new LinkedList<>();
-            products.addAll(searchProductByKeywords(searchWords));
-            products.addAll(searchProductByCategory(searchWords));
-            products.addAll(searchProductByName(searchWords));
-            Set<ProductInfo> set = new HashSet<>(products.size());
-            products.removeIf(p -> !set.add(p));
-            return products;
+        List<ProductInfo> products = new LinkedList<>();
+        products.addAll(searchProductByKeywords(searchWords));
+        products.addAll(searchProductByCategory(searchWords));
+        products.addAll(searchProductByName(searchWords));
+        Set<ProductInfo> set = new HashSet<>(products.size());
+        products.removeIf(p -> !set.add(p));
+        return products;
     }
 
     public List<ProductInfo> filterByPriceRange(int minPrice, int maxPrice) {
@@ -195,7 +197,7 @@ public class Market {
     }
 
     private double calculatePriceOfBasket(int storeId, BasketInfo basketInfo, UserInfo userInfo,
-                                          List<String> coupons) throws PurchaseExceedsPolicyException{//add date of birth
+                                          List<String> coupons) throws PurchaseExceedsPolicyException {//add date of birth
         return Optional.of(storeRepository.getStore(storeId)).orElseThrow(
                 () -> new RuntimeException("Store with id " + storeId + " does not exist")
         ).calculatePriceOfBasket(basketInfo, userInfo, coupons);
@@ -291,12 +293,11 @@ public class Market {
     }
 
 
-
     /**
      * <H1>Conditions</H1>
      */
 
-    public void setPurchasePolicyCondition(int storeId, int userId, int conditionId) throws NoPermissionException{
+    public void setPurchasePolicyCondition(int storeId, int userId, int conditionId) throws NoPermissionException {
         storeRepository.getStore(storeId).setPurchasePolicyCondition(userId, conditionId);
     }
 
@@ -319,19 +320,19 @@ public class Market {
         return storeRepository.getStore(storeId).addIMPLYCondition(userId, condition1, condition2);
     }
 
-    public int addStorePriceCondition(int storeId, int userId, double lowerBound, double upperBound) throws NoPermissionException{
+    public int addStorePriceCondition(int storeId, int userId, double lowerBound, double upperBound) throws NoPermissionException {
         return storeRepository.getStore(storeId).addStorePriceCondition(userId, lowerBound, upperBound);
     }
 
-    public int addStorePriceCondition(int storeId, int userId, double lowerBound) throws NoPermissionException{
+    public int addStorePriceCondition(int storeId, int userId, double lowerBound) throws NoPermissionException {
         return storeRepository.getStore(storeId).addStorePriceCondition(userId, lowerBound);
     }
 
-    public int addStoreQuantityCondition(int storeId, int userId, int lowerBound, int upperBound) throws NoPermissionException{
+    public int addStoreQuantityCondition(int storeId, int userId, int lowerBound, int upperBound) throws NoPermissionException {
         return storeRepository.getStore(storeId).addStoreQuantityCondition(userId, lowerBound, upperBound);
     }
 
-    public int addStoreQuantityCondition(int storeId, int userId, int lowerBound) throws NoPermissionException{
+    public int addStoreQuantityCondition(int storeId, int userId, int lowerBound) throws NoPermissionException {
         return storeRepository.getStore(storeId).addStoreQuantityCondition(userId, lowerBound);
     }
 
@@ -514,4 +515,13 @@ public class Market {
     public void removeIndividualPermission(int userId, int managerId, int storeId, UserPermissions.IndividualPermission individualPermission) throws NoPermissionException, ChangePermissionException {
         storeRepository.getStore(storeId).removePermissionFromManager(userId, managerId, individualPermission);
     }
+
+    public DiscountAccumulationTreeInfo getDiscountAccumulationTree(int storeId, int userId) throws NoPermissionException {
+        return storeRepository.getStore(storeId).getDiscountAccumulationTree(userId);
+    }
+
+    public void deleteStoreAccumulationTree(int storeId, int userId) throws NoPermissionException {
+        storeRepository.getStore(storeId).deleteStoreAccumulationTree(userId);
+    }
+
 }
