@@ -1,4 +1,4 @@
 package BGU.Group13B.backend.storePackage.newDiscoutns;
 
-public record DiscountInfo(String stringRepresentation) {
+public record DiscountInfo(int discountId, String stringRepresentation) {
 }
